% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExCW.R
\name{ExCW}
\alias{ExCW}
\title{Matched Filter peak picking demo}
\usage{
ExCW(ppm = 20, snthresh = 10)
}
\arguments{
\item{ppm}{the ppm tolerance used to identify "interesting" mass traces}

\item{snthresh}{snthresh the signal to noise threshold used to confirm that something is a peak}
}
\value{
A plot moitorig the process
}
\description{
Show how xcms performs peak picking by using the CentWave algorithm.
Some inline text is presented in the terminal to orient the discussion
}
\examples{

}

