% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExGroup.R
\name{ExGroup}
\alias{ExGroup}
\title{Density based Grouping demo}
\usage{
ExGroup(bw = 6, mzwd = 0.5)
}
\arguments{
\item{bw}{bandwidth (standard deviation or half width at half maximum)
of gaussian smoothing kernel to apply to the peak density chromatogram}

\item{mzwd}{width of overlapping m/z slices to use for creating peak density
chromatograms and grouping peaks across samples}
}
\value{
a plot showing the progress grouping along the m/z scale
}
\description{
Density based Grouping demo
}
\examples{

}

