% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExMF.R
\name{ExMF}
\alias{ExMF}
\title{Matched Filter peak picking demo}
\usage{
ExMF(step = 0.5, snthresh = 10)
}
\arguments{
\item{step}{the step used to slice the m/z range}

\item{snthresh}{the signal to noise threshold used to confirm that something is a peak}
}
\value{
A plot moitorig the process
}
\description{
Show how xcms performs peak picking by using the matched filter algorithm.
Some inline text is presented in the terminal to orient the discussion
}
\examples{


}

