% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptw.R
\name{ExWarp}
\alias{ExWarp}
\title{ExWarp}
\usage{
ExWarp()
}
\value{
an interactive demo on runnin gon the console. A web based shiny version of the demo is also available
}
\description{
Demo showing how parametric time warping (ptw) can be used to perform a simple form of
retention time alignment
}
\examples{

}

