% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyExample.R
\name{shinyExample}
\alias{shinyExample}
\title{shiny based xcms demo}
\usage{
shinyExample(example)
}
\arguments{
\item{name}{the name of the app}
}
\value{
a standard shiny app
}
\description{
Depending on the argument, run a web based interactive app to demonstrate the use of xcms
}
\examples{
## show he list of all the available demos
shinyExample()

}

