% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewCW.R
\name{viewCW}
\alias{viewCW}
\title{Visualize the peaks - centWave}
\usage{
viewCW(ppm = 10, snthresh = 10)
}
\arguments{
\item{ppm}{the ppm tolerance used to identify "interesting" mass traces}

\item{snthresh}{snthresh the signal to noise threshold used to confirm that something is a peak}
}
\value{
the mapping of the peaks
}
\description{
This demo shows the distribution of the peaks in the rt/mz plane, when the paek picking is performed
with centWave. This is the terminal version of the shiny app with the same name.
}
\examples{

}

