% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewMF.R
\name{viewMF}
\alias{viewMF}
\title{viewMF}
\usage{
viewMF(step = 0.5, snthresh = 10)
}
\arguments{
\item{step}{the step used to generate the profile matrix}

\item{snthresh}{snthresh the signal to noise threshold used to confirm that something is a peak}
}
\value{
a plot showing the ditribution of the peaks in the mz/rt plane
}
\description{
This demo shows the distribution of the peaks in the rt/mz plane, when the paek picking is performed
with Matched Filter. This is the terminal version of the shiny app with the same name.
}
\examples{

}

